==> src/02_Sagemaker_Inference.ipynb <==
##

import boto3
import sagemaker
import pandas as pd

sess = sagemaker.Session()
bucket = sess.default_bucket()
role = sagemaker.get_execution_role()
region = boto3.Session().region_name
prefix = 'mlops/data'

## Test Data

test_x_path = 's3://sagemaker-us-east-1-752847213914/mlops/data/test/test_x.csv'
test_y_path = 's3://sagemaker-us-east-1-752847213914/mlops/data/test/test_y.csv'

test_y = pd.read_csv(test_y_path, header = None)
test_y.columns = ['y']

test_y.shape

## Model Prediction

from sagemaker.transformer import Transformer

transformer = Transformer(model_name='rf-scikit-2023-07-14-09-27-43-800',instance_count=1,instance_type='ml.m5.2xlarge')

#prediction = transformer.transform(test_x_path)

sklearn_transformer = sklearn_estimator.transformer(
    model_name='rf-scikit-2023-07-14-09-27-43-800',
    instance_count=1,
    instance_type='ml.m5.large',
    output_path=transformer_output_path
)

# cell 18
transformer_output_path = f"s3://{bucket}/{prefix}/transformer-output"

sklearn_transformer = Transformer(
    model_name='rf-scikit-2023-07-14-09-27-43-800',
    instance_count=1,
    instance_type='ml.m5.large',
    output_path=transformer_output_path
)

sklearn_transformer.transform(
    data=test_x_path,
    data_type='S3Prefix',
    content_type='text/csv'
)

# cell 19
!aws s3 cp $transformer_output_path/test_x.csv.out /tmp/predictions.csv

with open('/tmp/predictions.txt', 'r') as r:
    a = r.read()
prediction = [int(x) for x in a[1:-1].split(',')]
prediction_df = pd.DataFrame({'prediction':prediction})

prediction_df.shape, test_y.shape

# cell 21
pd.crosstab(index=test_y['y'].values, columns=prediction_df['prediction'].values, rownames=['actuals'], colnames=['predictions'])

### Data Analytics
