==> src/03_Sagemaker_Realtime_Inference.ipynb <==
# Batch Time and Real-time Prediction

# Importing the necessary library
import boto3
import sagemaker
import pandas as pd
from joblib import dump, load
import s3fs

# Initialising new sagemaker session as "sess".
sess = sagemaker.Session()
# Bucket variable is used for storing the location of the bucket
bucket = 'sagemaker-studio-009676737623-l4vs7j0o0ib'
# Assigning the prefix variable
prefix = 'mlops-level1-data'
# Check for necessary permission needed for training and deploying models. 
role = sagemaker.get_execution_role()
# To understand where this session is configured to operate.
region = boto3.Session().region_name

## Test Data

# Creating a string for the test path 
test_data_path = f's3://{bucket}/{prefix}/test_data.csv'

# Importing the dataset
test_data = pd.read_csv(test_data_path)
test_data.shape

# Treating missing values
test_data.isna().sum()
test_data.dropna(inplace = True)
test_data.shape

## Get Features
fs = s3fs.S3FileSystem() # Updated method name
filename = f's3://{bucket}/{prefix}/feature/feature.joblib'
with fs.open(filename, encoding='utf8') as fh:
    cols = load(fh)

## Get Encoder object
filename = f's3://{bucket}/{prefix}/feature/encoder.joblib'
with fs.open(filename, encoding='utf8') as fh:
    encoder = load(fh)

cols

processed_test_data = test_data[cols]
test_x = processed_test_data.drop('Activity', axis = 1)
test_y= processed_test_data[['Activity']]
test_x.shape, test_y.shape


## Save Processed Test to S3

test_x_path = f"s3://{bucket}/{prefix}/prediction/test/test_x.csv"
test_y_path =f"s3://{bucket}/{prefix}/prediction/test/test_y.csv"
test_x.to_csv(test_x_path, index=False, header=False)
test_y.to_csv(test_y_path, index=False, header=False)


# Real time Prediction

client = boto3.client('sagemaker-runtime')

endpoint_name = 'HumanActivity-InferenceEndpoint-2023-09-25-0546'

from sagemaker.predictor import Predictor
from sagemaker.serializers import CSVSerializer, JSONSerializer
from sagemaker.deserializers import CSVDeserializer , JSONDeserializer

predictor_csv = Predictor(endpoint_name=endpoint_name,
                      serializer=CSVSerializer(),
                     deserializer=CSVDeserializer())

prediction = predictor_csv.predict(test_x.sample(10))

### Check Data Capture

data_capture = 's3://sagemaker-studio-009676737623-l4vs7j0o0ib/mlops-level1-data/datacapture/HumanActivity-InferenceEndpoint-2023-09-25-0546/AllTraffic/2023/09/25/05/52-24-109-c21b0608-1627-419c-b2b4-cb46d0bc9471.jsonl'

json_df = pd.read_json(data_capture,lines=True)

json_df

input_data = json_df['captureData'][0]

input_data
